[step3] refactor: 패키지 추가 및 정리 1차
