+++ src/main/java/lotto/numbers/AllNumbers.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.numbers;
 
 import java.util.List;
 import java.util.stream.Collectors;
